+++ docs/fr_FR/Changelog.md
@@ -1,5 +1,7 @@
 # ChangeLog
 
+- Correction: Parser: 'Dev announce' ignoré que si FW version < 3.22 (2592).
+
 ## 241220-BETA-1
 
 -   Nouveau: Support Legrand wireless switch (2736).